Refactor task command and move cache-related functions to utils
